% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.offsets}
\alias{get.offsets}
\title{Gets depths from data frame containing profile info.}
\usage{
get.offsets(data)
}
\arguments{
\item{data}{Data frame returned from \link{load.ts}.}
}
\value{
A numeric vector of depth values. Should be the ncol(data) -
1 in length as the first column contains date/time data.
}
\description{
Extracts the depth information from a data frame
containing multi-depth observation data. Relies on the format
of the header to get information and may fail if your file
format is incorrect.
}
\examples{
#Get the path for the package example file included
exampleFilePath <- system.file('extdata', 'Sparkling.wtr', package="rLakeAnalyzer")

#Load
sparkling.temp = load.ts(exampleFilePath)

#get the lake depths associated with each column
depths = get.offsets(sparkling.temp)

print(depths)
}
\author{
Luke Winslow
}
\seealso{
\link{load.ts}
}

